rework pub and use
