+++ src/lib.rs
@@ -7,22 +7,23 @@
 #![allow(clippy::module_name_repetitions)]
 
 pub mod args;
-pub mod config;
+mod config;
 pub mod pg;
-pub mod source;
+mod source;
 pub mod srv;
-pub mod utils;
+mod utils;
 
-pub use crate::utils::Error;
-pub use crate::utils::Result;
+#[cfg(test)]
+#[path = "utils/test_utils.rs"]
+mod test_utils;
 
 // test_utils is used from tests in other modules, and it uses this crate's object.
 // Must make it accessible as carte::Env from both places when testing.
 #[cfg(test)]
-pub use crate::args::environment::Env;
-#[cfg(test)]
-#[path = "utils/test_utils.rs"]
-mod test_utils;
+pub use crate::args::Env;
+pub use crate::config::{read_config, Config};
+pub use crate::source::{IdResolver, Source, Sources, Xyz};
+pub use crate::utils::{Error, Result};
 
 // Ensure README.md contains valid code
 #[cfg(doctest)]
