+++ src/pg/utils.rs
@@ -1,14 +1,14 @@
-use crate::source::{UrlQuery, Xyz};
-use crate::utils::InfoMap;
-use actix_http::header::HeaderValue;
-use actix_web::http::Uri;
+use std::collections::HashMap;
+
 use itertools::Itertools;
 use postgis::{ewkb, LineString, Point, Polygon};
 use postgres::types::Json;
 use semver::Version;
 use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use tilejson::{tilejson, Bounds, TileJSON, VectorLayer};
+
+use crate::source::{UrlQuery, Xyz};
+use crate::utils::InfoMap;
 
 #[must_use]
 pub fn json_to_hashmap(value: &serde_json::Value) -> InfoMap<String> {
@@ -53,14 +53,6 @@ pub fn polygon_to_bbox(polygon: &ewkb::Polygon) -> Option<Bounds> {
     })
 }
 
-pub fn parse_x_rewrite_url(header: &HeaderValue) -> Option<String> {
-    header
-        .to_str()
-        .ok()
-        .and_then(|header| header.parse::<Uri>().ok())
-        .map(|uri| uri.path().to_owned())
-}
-
 #[must_use]
 pub fn create_tilejson(
     name: String,
@@ -82,12 +74,6 @@ pub fn create_tilejson(
     // TODO: consider removing - this is not needed per TileJSON spec
     tilejson.set_missing_defaults();
     tilejson
-}
-
-#[must_use]
-pub fn is_valid_zoom(zoom: i32, minzoom: Option<u8>, maxzoom: Option<u8>) -> bool {
-    minzoom.map_or(true, |minzoom| zoom >= minzoom.into())
-        && maxzoom.map_or(true, |maxzoom| zoom <= maxzoom.into())
 }
 
 pub type Result<T> = std::result::Result<T, PgError>;
