+++ src/srv/mod.rs
@@ -1,2 +1,5 @@
-pub mod config;
-pub mod server;
+mod config;
+mod server;
+
+pub use config::{SrvConfig, KEEP_ALIVE_DEFAULT, LISTEN_ADDRESSES_DEFAULT};
+pub use server::{new_server, router, AppState, IndexEntry, RESERVED_KEYWORDS};
