+++ src/utils/mod.rs
@@ -1,5 +1,5 @@
-pub mod one_or_many;
-pub mod utilities;
+mod one_or_many;
+mod utilities;
 
-pub use one_or_many::*;
+pub use one_or_many::OneOrMany;
 pub use utilities::*;
